chore: update deps, add missing deps, remove unused deps
* Some deps were not being used (maybe from older
versions of this app?) so they were removed, including
ts-node (we compile and run js now) and node-fetch (we
use native fetch in node now), along with their associated
types packages.
* some deps were not in the package.json, which means
the app was using the transitive dependency instead of
declaring the dependency. Best practice is to be explicit
if the dependency is being used directly.
* also added the node version in package.json to
the engines section, mostly to indicate the minimum
node version that should be used.