chore: update dependencies, tracing setup, and HTTP attributes (#14)
* chore: update deps, add missing deps, remove unused deps

* Some deps were not being used (maybe from older
versions of this app?) so they were removed, including
ts-node (we compile and run js now) and node-fetch (we
use native fetch in node now), along with their associated
types packages.
* some deps were not in the package.json, which means
the app was using the transitive dependency instead of
declaring the dependency. Best practice is to be explicit
if the dependency is being used directly.
* also added the node version in package.json to
the engines section, mostly to indicate the minimum
node version that should be used.

* update for instr changes and best practices

* Instead of importing all of otel api, only
import specific modules that are needed.
* instr-fs is now disabled by default in
auto-instrumentations.
* add graceful shutdown for otel sdk

* update to stable http attributes

* add in dual-emit for http attributes

* http/dup emits both old and new (stable)
http attributes from semantic conventions.
This is to help migrate all services, SLOs,
queries, triggers, etc to use the new
attributes. Once all instrumentations use
new attributes, and all queries and triggers
are updated to look for new attributes,
this can change to only "http", and the
dual-emit option will eventually be removed
from otel-js.

* should these be v2.0 now? shrug

* app is in 10114 not 8080

* mostly automated linting from my IDE

* dont want these map files getting pushed