chore(instr-branch): update dependencies, tracing setup, and HTTP attributes for instrumented branch (#15)
* mostly automated linting from my IDE

* dont want these map files getting pushed

* app is in 10114 not 8080

* should these be v2.0 now? shrug

* chore: update deps, add missing deps, remove unused deps

* Some deps were not being used (maybe from older
versions of this app?) so they were removed, including
ts-node (we compile and run js now) and node-fetch (we
use native fetch in node now), along with their associated
types packages.
* some deps were not in the package.json, which means
the app was using the transitive dependency instead of
declaring the dependency. Best practice is to be explicit
if the dependency is being used directly.
* also added the node version in package.json to
the engines section, mostly to indicate the minimum
node version that should be used.

* more version updates

* add permission to allow stop

* update docker-compose to match main

* not sure why the ports had to be different
but running this it seems to work just fine
so I updated it to use that.

* update for instr changes and best practices

* Instead of importing all of otel api, only
import specific modules that are needed.
* instr-fs is now disabled by default in
auto-instrumentations.
* add graceful shutdown for otel sdk

* typos and fix comment syntax in .env

* update to newer attrs

* add in dual-emit for http attrs

* http/dup emits both old and new (stable)
http attributes from semantic conventions.
This is to help migrate all services, SLOs,
queries, triggers, etc to use the new
attributes. Once all instrumentations use
new attributes, and all queries and triggers
are updated to look for new attributes,
this can change to only "http", and the
dual-emit option will eventually be removed
from otel-js.

* only copy typescript files in docker build

* this is mostly for local dev, when there are
still javascript files from previous builds
* also generally best practice to only copy
as much as needed and no more.

* add direct deps for instrumenting

* update semconv imports for stability

* Service version is a stable attribute
so that can be imported as-is.
* Latest guidance for experimental semconv
is to keep a local copy of definitions to avoid
breakage in the incubating entrypoint. It is
no longer recommended to use incubating
entrypoint directly.
* Local semconv file added using definitions
from latest version of otel-js semantic
conventions experimental_attributes,
along with link to explanation.

* update resource class to new function

* sdk v2 breaking change included change
from using Resource class to new function
resourceFromAttributes.

* Apply suggestions from code review

Co-authored-by: Mike Goldsmith <goldsmith.mike@gmail.com>

* remove local semconv for now

---------

Co-authored-by: Mike Goldsmith <goldsmith.mike@gmail.com>